Move atomic, make_connection_plus_from to sqlite.connection
